Update channel name and export isInCI in devicefarm lane
This mirrors the logic of how a channel name is created for the test
suite app in `test-suite/publish.sh`. Instead of just using git_branch it converts
it to lowercase and removes all invalid characters so that there is no
discrepancy between that and the published version. I also export a
isInCI env variable in Test config which we will use to skip tests that
require interaction (e.g. signing in).